feat(web): split Security, Appearance, Sessions and Plan out of Profile and Preferences
Four new settings panes under `features/workspace/settings/tabs/`, each the
same pure-view + container split the existing panes use:

- `security-tab.tsx`: two-factor authentication moved out of Profile, plus a
  "Sign out other devices" row (`supabase.auth.signOut({ scope: 'others' })`).
  `FactorRow` and `totpQrSrc` move here with the section that renders them.
- `appearance-tab.tsx`: theme, conversation density and wallpaper, out of
  Preferences.
- `sessions-tab.tsx`: browser notifications and sounds, out of Preferences.
- `plan-tab.tsx`: the account's subscription, credits and billing portal. A
  second mount of the account page's `BillingTab`, inside the same
  `BillingAccountProvider` and behind the same `account.write` gate; a member
  without it gets a link to the account page instead of buttons that fail.

Profile keeps picture, email, name, organizations and account deletion.
Preferences keeps language (first) and keyboard shortcuts.

`settings-tab-header.test.ts` and `tab-content-width.test.ts` classify the
four new files; the profile and preferences tests move their two-factor and
appearance cases to the panes that own them now.
