+++ apps/web/src/features/workspace/settings/tabs/profile-tab.test.tsx
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'bun:test';
 import { renderToStaticMarkup } from 'react-dom/server';
-import { FactorRow, ProfileTabView, totpQrSrc } from './profile-tab';
+import { ProfileTabView } from './profile-tab';
 
 /** Section titles in document order, read from the h2s the pane emits — the
  *  page heading (`SettingsTabHeader`) plus each section label. Row labels are
@@ -14,18 +14,12 @@ const html = () => renderToStaticMarkup(<ProfileTabView />);
 
 describe('ProfileTabView', () => {
   test('renders the pane heading and each section label, in order', () => {
-    expect(headings(html())).toEqual(['Profile', 'Security', 'Danger zone']);
+    expect(headings(html())).toEqual(['Profile', 'Danger zone']);
   });
 
   test('renders every setting row, in order', () => {
     const out = html();
-    const rows = [
-      'Profile picture',
-      'Email',
-      'Name',
-      'Two-factor authentication',
-      'Delete account',
-    ];
+    const rows = ['Profile picture', 'Email', 'Name', 'Delete account'];
     const positions = rows.map((label) => out.indexOf(`>${label}<`));
     expect(positions.some((p) => p < 0)).toBe(false);
     expect([...positions].sort((a, b) => a - b)).toEqual(positions);
@@ -56,13 +50,8 @@ describe('ProfileTabView', () => {
       expect(out).toContain('>Acme<');
     });
 
-    test('sits under the identity rows and above Security', () => {
-      expect(headings(withAccounts())).toEqual([
-        'Profile',
-        'Organizations',
-        'Security',
-        'Danger zone',
-      ]);
+    test('sits under the identity rows and above Danger zone', () => {
+      expect(headings(withAccounts())).toEqual(['Profile', 'Organizations', 'Danger zone']);
       expect(withAccounts().indexOf('>Email<')).toBeLessThan(
         withAccounts().indexOf('Organizations'),
       );
@@ -72,7 +61,7 @@ describe('ProfileTabView', () => {
      *  accounts query has not answered. This is why the heading assertion at
      *  the top of this file still reads three headings. */
     test('adds nothing to the pane when the list is unknown', () => {
-      expect(headings(html())).toEqual(['Profile', 'Security', 'Danger zone']);
+      expect(headings(html())).toEqual(['Profile', 'Danger zone']);
     });
   });
 
@@ -118,93 +107,14 @@ describe('ProfileTabView', () => {
 });
 
 /**
- * `factorsError` (fed by `useMfa()`'s `factorsQuery.isError`) must render a
- * real error state with a retry action instead of falling through to the
- * "No second factor enrolled" empty-state copy — that copy is a factual
- * claim that the factor list came back empty, not that it failed to load.
+ * Two-factor authentication is not on this pane any more — it is the Security
+ * tab (`security-tab.tsx`, pinned by `security-tab.test.tsx`). Asserted absent
+ * so the section cannot quietly grow back here as a second copy.
  */
-describe('ProfileTabView — two-factor error state', () => {
-  test('a failed factors fetch shows an error, not the empty-state banner', () => {
-    const out = renderToStaticMarkup(<ProfileTabView factorsError />);
-    expect(out).toContain('load your authenticator apps');
-    expect(out).toContain('>Retry<');
+describe('ProfileTabView — no security section', () => {
+  test('renders no two-factor row and no factor banner', () => {
+    const out = html();
+    expect(out).not.toContain('Two-factor authentication');
     expect(out).not.toContain('No second factor enrolled');
-  });
-
-  test('loading takes priority over the error state', () => {
-    const out = renderToStaticMarkup(<ProfileTabView factorsLoading factorsError />);
-    expect(out).not.toContain('load your authenticator apps');
-  });
-
-  test('no error by default — the empty-state banner renders instead', () => {
-    expect(html()).toContain('No second factor enrolled');
-  });
-});
-
-/**
- * The other two answers the factor list can give. Loading is a shape-matched
- * skeleton, not a blank gap; a populated list is its own answer, so neither
- * banner may appear next to it. Added with the error state above — before it,
- * `{!factorsLoading && !factorsError ? factors.map(…) : null}` had no else
- * branch at all, so two of the four states rendered literally nothing.
- */
-describe('ProfileTabView — two-factor list states', () => {
-  test('an in-flight factor list shows a skeleton, not a blank gap', () => {
-    const out = renderToStaticMarkup(<ProfileTabView factorsLoading />);
-    expect(out).toContain('animate-pulse');
-    expect(out).not.toContain('No second factor enrolled');
-  });
-
-  test('an enrolled factor renders as a row, with neither banner', () => {
-    const out = renderToStaticMarkup(
-      <ProfileTabView
-        factors={[{ id: 'f1', friendly_name: 'My phone', factor_type: 'totp', status: 'verified' }]}
-      />,
-    );
-    expect(out).toContain('My phone');
-    expect(out).not.toContain('No second factor enrolled');
-    expect(out).not.toContain('load your authenticator apps');
-  });
-});
-
-// Ported from the deleted `features/accounts/settings/security-tab.test.tsx`
-// (Task 10) — `FactorRow` and `totpQrSrc` moved into `profile-tab.tsx`, this
-// tab being their only remaining consumer once the legacy user-settings
-// modal and `security-tab.tsx` were retired.
-describe('totpQrSrc', () => {
-  test('passes a data URL through untouched', () => {
-    const url = 'data:image/svg+xml;utf8,%3Csvg%3E%3C/svg%3E';
-    expect(totpQrSrc(url)).toBe(url);
-  });
-
-  test('wraps raw SVG into an encoded data URL', () => {
-    const out = totpQrSrc('<svg xmlns="http://www.w3.org/2000/svg"></svg>');
-    expect(out.startsWith('data:image/svg+xml;utf8,')).toBe(true);
-    expect(out).toContain('%3Csvg');
-  });
-});
-
-describe('FactorRow', () => {
-  test('verified authenticator renders name, type, and verified badge', () => {
-    const factorHtml = renderToStaticMarkup(
-      <FactorRow
-        factor={{ id: 'f1', friendly_name: 'My phone', factor_type: 'totp', status: 'verified' }}
-        onRemove={() => {}}
-      />,
-    );
-    expect(factorHtml).toContain('My phone');
-    expect(factorHtml).toContain('verified');
-    expect(factorHtml).toContain('Remove factor');
-  });
-
-  test('unnamed totp factor falls back to "Authenticator app"', () => {
-    const factorHtml = renderToStaticMarkup(
-      <FactorRow
-        factor={{ id: 'f2', factor_type: 'totp', status: 'unverified' }}
-        onRemove={() => {}}
-      />,
-    );
-    expect(factorHtml).toContain('Authenticator app');
-    expect(factorHtml).toContain('unverified');
   });
 });
